Extracted round robin load balancing strategy in a separate file
