+++ test/internal/round-robin-load-balancing-strategy.test.js
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {RoundRobinLoadBalancingStrategy} from '../../src/v1/internal/load-balancing-strategies';
+import RoundRobinLoadBalancingStrategy from '../../src/v1/internal/round-robin-load-balancing-strategy';
 
 describe('RoundRobinLoadBalancingStrategy', () => {
